# Shadow volume culling and tighter shadow caster culling