--- a/core/error/error_macros.h
+++ b/core/error/error_macros.h
@@ -812,4 +812,14 @@ void _err_flush_stdout();
 #define DEV_ASSERT(m_cond)
 #endif
 
+#ifdef DEV_ENABLED
+#define DEV_CHECK_ONCE(m_cond)                                                   \
+	if (unlikely(!(m_cond))) {                                                   \
+		ERR_PRINT_ONCE("DEV_CHECK_ONCE failed  \"" _STR(m_cond) "\" is false."); \
+	} else                                                                       \
+		((void)0)
+#else
+#define DEV_CHECK_ONCE(m_cond)
+#endif
+
 #endif // ERROR_MACROS_H