… servers/rendering/rendering_light_culler.cpp