@clayjohn 2023-11-13
I tested quickly on a high complexity scene and it spammed the following error:
```