@lawnjelly 2023-11-14
> I tested quickly on a high complexity scene and it spammed the following error:

I'll take a look. This may be due to some slight differences in cameras from 3.x (bear in mind this is a straight port and has had much more testing on 3.x). :+1: