@clayjohn 2023-11-14
Quick debugging check, changing MAX_PLANES to 32 makes the error go away.

The scene has multiple omnilights and a directionallight3D. The error spams while multiple omnilights are within the camera frustum. The error continues when the DirectionalLight3D is made not visible. 

I'll try to make a reproduction project