@lawnjelly 2023-11-14
Also worth checking is if setting `MAX_PLANES` to 17 makes it go away.

I've just pushed a slightly updated version with the lookup table showing all 8 entries (there was a strange half finished bit for only writing 7 entries on purpose, but I wrote this 3 years ago so can only try and remember why I did that, maybe it was a bug).

I'm not sure this would have created your bug but it is just possible, so try the latest version as well.