@clayjohn 2023-11-14
Okay, I've prepared an MRP that reliably reproduces the issue. 
[clip-plane-mrp.zip](https://github.com/godotengine/godot/files/13349743/clip-plane-mrp.zip)