@lawnjelly 2023-11-14
Fantastic! Thanks so much for spending time on this. I can immediately see that light culler contains some garbage values so some kind of corruption is going on.