@lawnjelly 2023-11-20
This should be fixed now.

* Preparing the directional lights has been moved outside the multithreaded code, only the culling is performed there now, which should be thread safe.
* I also fixed it to work with multiple directional lights. These are prepared in advance, so the `cull_directional_light()` can chop and change between the different directional lights (this is the pattern currently used in the code).