@Calinou 2023-11-20
Tested locally, it works as expected. Visuals look correct too from my testing in various demo projects.

## Benchmark on [tps-demo](https://github.com/godotengine/tps-demo)

**OS:** Fedora 38
**CPU:** Intel Core i9-13900K
**GPU:** GeForce RTX 4090 (NVIDIA 535.113.01)

The project is modified to disable V-Sync. The FPS reported is the highest FPS attained over a period of 10 seconds after loading the level. Unfortunately, unlike the `3.x` branch, I can't notice any performance improvement with this PR. This also applies to other demos like Truck Town and 3D platformer which I've tested as well.

Type | Before | After
-|-|-
4K Maximum Forward+ | 232 | 234
4K Minimum Forward+ | 961 | 962
720p Maximum Forward+[^1] | 579 | 568
720p Minimum Forward+[^1] | 1044 | 1033
4K Minimum **Compatibility**[^1] | 471 | 472
720p Minimum **Compatibility**[^1] | 466 | 476

[^1]: CPU is the bottleneck, with GPU sitting at less than 50% utilization.