@lawnjelly 2023-11-20
> Unfortunately, unlike the 3.x branch, I can't notice any performance improvement with this PR.

Oh I must have made a booboo in the fixup then, as there should be an improvement. I'll do a debug to check it's working correctly (I don't have many test projects in 4.x).