@clayjohn 2023-12-01
I tested again on the same scene and can confirm that the errors are gone. 

In a section of the level that is mostly outdoors, I noticed a decrease in draw calls by about 25% (2500 -> 1900). 

However, in an indoor section with multiple omnilights, I noticed an increase in the number of draw calls (14,300 -> 14,700). My guess is that something is forcing the omnilights to update every frame instead of just when changed. I'll try to isolate the case and confirm the theory