@lawnjelly 2023-12-02
> However, in an indoor section with multiple omnilights, I noticed an increase in the number of draw calls (14,300 -> 14,700). My guess is that something is forcing the omnilights to update every frame instead of just when changed. I'll try to isolate the case and confirm the theory

I'll have a look too in the mean time and see if I can work out what is going on.