@lawnjelly 2024-01-30
After testing the large MRP @clayjohn mentioned above, my results are as follows:

Drawcalls (tight culling first, legacy second):
```
Outdoor 2530, 2873
Indoor 19990, 21440
Panoramic 3487, 4713
Free 1977, 7191
Character 1972, 2315
```
So improvement in all cases.

I haven't figured out yet how to move the indoor camera, these are the default views.

BTW I tested by switching on the `LIGHT_CULLER_DEBUG_FLASH` define, and changing the frequency to 64 (I got very low frame rate on my PC).
Then recorded the drawcalls given in the game.
Also I set `LIGHT_CULLER_DEBUG_LOGGING`, it tells you when it flashes the culling on and off.
Also I realised there was an out of date variable name in the debugging defined section, had to swap an `active` to `light_culling_active` (due to a variable rename) to get it to compile. I'll fix that in the PR and repush.