@mrjustaguy 2024-03-19
Tighter Shadow caster culling causes the shadows to disappear in some cases...

Run this test scene, and press home a handful of times to rotate the lights one increment at a time (end to rotate in reverse)
With tighter shadow caster culling enabled, shadows will disappear for a frame a few rotations in, only to pop back in, but this is not the case with it disabled
[Shadows.zip](https://github.com/godotengine/godot/files/14655276/Shadows.zip)