@lawnjelly 2024-03-19
> Tighter Shadow caster culling causes the shadows to disappear in some cases...

This could be a bug, usually better reported in an issue. I'll try out the project. :+1: