@mrjustaguy 2024-03-20
Sorry about that, was kind of in a rush yesterday and it was a bit late, The PR happened to be the 1st thing I was thinking about when I observed and quickly pinned down the issue.

What doesn't make sense is that it only disappears for 1 frame, only to reaper the next despite the fact there should be no update to lights that next frame, as they're static... So somehow they're dirtied one frame after the rotation, probably by the offending model (Female26 for me) and this has nothing to do with skinning as yea, a cube of roughly the AABB of the dummy also gets the same issue.

I'll open an issue now..